btsk-86: 기본폴더 삭제 불가 로직 및 폴더 조회 api 로직 변경 data.sql 삽입.

+++ src/main/java/kr/it/pullit/modules/commonfolder/service/FolderFacadeImpl.java
@@ -2,6 +2,9 @@ package kr.it.pullit.modules.commonfolder.service;
 
 import kr.it.pullit.modules.commonfolder.api.CommonFolderPublicApi;
 import kr.it.pullit.modules.commonfolder.api.FolderFacade;
+import kr.it.pullit.modules.commonfolder.domain.entity.CommonFolder;
+import kr.it.pullit.modules.commonfolder.exception.CommonFolderErrorCode;
+import kr.it.pullit.modules.commonfolder.exception.InvalidFolderOperationException;
 import kr.it.pullit.modules.questionset.api.QuestionSetPublicApi;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -23,6 +26,9 @@ public class FolderFacadeImpl implements FolderFacade {
   @Override
   @Transactional
   public void deleteFolderAndContents(Long folderId) {
+    if (folderId.equals(CommonFolder.DEFAULT_FOLDER_ID)) {
+      throw new InvalidFolderOperationException(CommonFolderErrorCode.CANNOT_DELETE_DEFAULT_FOLDER);
+    }
     questionSetPublicApi.deleteAllByFolderId(folderId);
     commonFolderPublicApi.deleteFolder(folderId);
   }
